% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_tissue_functions.R
\name{aggregate_by_tissue}
\alias{aggregate_by_tissue}
\title{Aggregate SNP-level posteriors into tissue groups}
\usage{
aggregate_by_tissue(em_matrix, snp_info, track_info)
}
\arguments{
\item{em_matrix}{Matrix of SNP-by-track posteriors}

\item{snp_info}{Summary stats with SNP identifiers}

\item{track_info}{Table linking tracks to tissues}
}
\value{
A data.table with SNP-by-tissue enrichment scores
}
\description{
Aggregate SNP-level posteriors into tissue groups
}
