% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_tissue_functions.R
\name{build_track_metadata}
\alias{build_track_metadata}
\title{Build track-to-tissue metadata from sample IDs}
\usage{
build_track_metadata(sumstats, metadata, sampleID)
}
\arguments{
\item{sumstats}{Summary stats with overlap columns}

\item{metadata}{Metadata table linking tracks to tissues}

\item{sampleID}{Sample ID table}
}
\value{
A data.table linking tracks to tissue groupings
}
\description{
Build track-to-tissue metadata from sample IDs
}
