% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_tissue_functions.R
\name{compute_background_matrix}
\alias{compute_background_matrix}
\title{Compute background expectation from 1KG chr1 and tracks}
\usage{
compute_background_matrix(dir_epi, tracks, exons)
}
\arguments{
\item{dir_epi}{Directory containing 1KG .bim file}

\item{tracks}{Combined epigenomic tracks}

\item{exons}{A data.table of exon coordinates}
}
\value{
A named numeric vector of background expectations
}
\description{
Compute background expectation from 1KG chr1 and tracks
}
