% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_tissue_functions.R
\name{compute_track_overlaps}
\alias{compute_track_overlaps}
\title{Compute SNP-track overlaps from BED data}
\usage{
compute_track_overlaps(sumstats, tracks)
}
\arguments{
\item{sumstats}{A data.table of summary statistics}

\item{tracks}{A data.table of epigenomic tracks}
}
\value{
A data.table with overlap annotations
}
\description{
Compute SNP-track overlaps from BED data
}
