% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concensus_run.R
\name{concensus_run}
\alias{concensus_run}
\title{Retrieve concensus run}
\usage{
concensus_run(runs, V_trait, V_tissue)
}
\arguments{
\item{runs}{bNMF runs}

\item{V_trait}{Trait matrix}

\item{V_tissue}{Tissue matrix}
}
\description{
Retrieve concensus run
}
