% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_test_data.R
\name{download_test_data}
\alias{download_test_data}
\title{Download example test data for jpepR}
\usage{
download_test_data(dest_dir = "data-raw", overwrite = FALSE)
}
\arguments{
\item{dest_dir}{Local directory to save downloaded files (default: "data-raw")}

\item{overwrite}{Logical. If TRUE, overwrite existing files. Default is FALSE.}
}
\value{
Invisibly returns the path to the downloaded folder
}
\description{
Downloads test files needed to run the vignette and example workflows.
}
