% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_tissue_functions.R
\name{load_and_filter_tracks}
\alias{load_and_filter_tracks}
\title{Load and combine epigenomic BED files, filter outliers}
\usage{
load_and_filter_tracks(dir_epi, pattern = ".bed", outlier_length = 5e+05)
}
\arguments{
\item{dir_epi}{Directory containing BED files}

\item{pattern}{File pattern to match BED files}

\item{outlier_length}{Length threshold for excluding outlier peaks}
}
\value{
A cleaned data.table of combined tracks
}
\description{
Load and combine epigenomic BED files, filter outliers
}
