% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_tissue_functions.R
\name{load_custom_metadata}
\alias{load_custom_metadata}
\title{Load metadata file linking tracks to tissues}
\usage{
load_custom_metadata(dir_epi, pattern = "metadata")
}
\arguments{
\item{dir_epi}{Directory containing metadata file}

\item{pattern}{Character pattern to match filename (default: "metadata")}
}
\value{
A data.table of metadata
}
\description{
Load metadata file linking tracks to tissues
}
