% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_tissue_functions.R
\name{load_custom_sample_ids}
\alias{load_custom_sample_ids}
\title{Load sample IDs associated with epigenomic tracks}
\usage{
load_custom_sample_ids(dir_epi)
}
\arguments{
\item{dir_epi}{Directory containing sample ID file}
}
\value{
A data.table of sample IDs
}
\description{
Load sample IDs associated with epigenomic tracks
}
