% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_tissue_functions.R
\name{load_summary_stats}
\alias{load_summary_stats}
\title{Load focal summary statistics}
\usage{
load_summary_stats(trait, output_dir)
}
\arguments{
\item{trait}{Name of the focal trait}

\item{output_dir}{Directory containing SQLite databases}
}
\value{
A data.table of summary statistics
}
\description{
Loads summary statistics from an SQLite database for a given trait.
}
