% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_vtissue.R
\name{make_vtissue}
\alias{make_vtissue}
\title{Construct V_tissue matrix}
\usage{
make_vtissue(
  focal_trait,
  output_dir,
  epi_dir = "default",
  pip_threshold = 0.01
)
}
\arguments{
\item{focal_trait}{Trait name}

\item{output_dir}{Output directory}

\item{epi_dir}{Path to epigenomic BED files or "default"}

\item{pip_threshold}{Minimum PIP}
}
\description{
Construct V_tissue matrix
}
