% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_vtrait.R
\name{make_vtrait}
\alias{make_vtrait}
\title{Construct V_trait matrix}
\usage{
make_vtrait(
  focal_trait,
  aux_traits,
  output_dir,
  pip_threshold = 0.01,
  aux_pip_thres = 0.5,
  aux_nbr_thres = 1
)
}
\arguments{
\item{focal_trait}{Name of focal trait}

\item{aux_traits}{Vector of auxiliary trait names}

\item{output_dir}{Output directory}

\item{pip_threshold}{Minimum PIP threshold}

\item{aux_pip_thres}{Min PIP for aux trait}

\item{aux_nbr_thres}{Min number of traits with strong PIP per SNP}
}
\description{
Construct V_trait matrix
}
