% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_tissue_functions.R
\name{normalize_cpt}
\alias{normalize_cpt}
\title{Normalize conditional probability table (CPT)}
\usage{
normalize_cpt(mat)
}
\arguments{
\item{mat}{Matrix of SNP-track binary overlaps}
}
\value{
A normalized CPT matrix (row-normalized)
}
\description{
Normalize conditional probability table (CPT)
}
