% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_tissue_functions.R
\name{prepare_em_input}
\alias{prepare_em_input}
\title{Prepare EM input matrix from filtered overlaps}
\usage{
prepare_em_input(sumstats_epi, track_info)
}
\arguments{
\item{sumstats_epi}{Annotated SNP-track matrix}

\item{track_info}{Track metadata table}
}
\value{
A binary matrix ready for EM
}
\description{
Prepare EM input matrix from filtered overlaps
}
