% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_tissue_functions.R
\name{run_em_algorithm}
\alias{run_em_algorithm}
\title{Run EM algorithm to estimate SNP-to-track posteriors}
\usage{
run_em_algorithm(
  cpt,
  pip,
  theta0,
  bckg_expectation,
  tol = 1e-08,
  maxiter = 200
)
}
\arguments{
\item{cpt}{Normalized conditional probability table}

\item{pip}{Vector of PIPs for each SNP}

\item{theta0}{Initial prior values}

\item{bckg_expectation}{Vector of background expectations}

\item{tol}{Convergence tolerance}

\item{maxiter}{Maximum number of iterations}
}
\value{
A list with final posterior matrix and iteration count
}
\description{
Run EM algorithm to estimate SNP-to-track posteriors
}
