% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_finemap.R
\name{validate_finemap}
\alias{validate_finemap}
\title{Validate a fine-mapped file format}
\usage{
validate_finemap(file)
}
\arguments{
\item{file}{Path to a fine-mapped file (e.g. .tsv or .tsv.gz)}
}
\value{
Invisibly returns TRUE if valid, otherwise throws an error
}
\description{
Ensures that the custom fine-mapped file contains required columns and that
\code{PIP} values are properly defined between 0 and 1.
}
