% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_ld_blocks.R
\name{validate_ld_blocks}
\alias{validate_ld_blocks}
\title{Validate custom LD blocks file or object}
\usage{
validate_ld_blocks(ld_blocks)
}
\arguments{
\item{ld_blocks}{A data.frame or data.table with LD blocks}
}
\value{
Invisibly returns TRUE if valid, otherwise throws an error
}
\description{
Ensures that the LD block object has the correct format and values.
}
