% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_vtissue.R
\name{make_vtissue}
\alias{make_vtissue}
\title{Construct V_tissue matrix from summary statistics and epigenomic tracks}
\usage{
make_vtissue(
  focal_trait,
  output_dir,
  epi_dir = "default",
  pip_threshold = 0.01
)
}
\arguments{
\item{focal_trait}{Name of the focal trait}

\item{output_dir}{Output directory}

\item{epi_dir}{Path to custom epigenomic BED folder or "default" for built-in EpiMap}

\item{pip_threshold}{Minimum PIP threshold to retain SNPs (default = 0.01)}
}
\value{
Path to saved V_tissues file
}
\description{
Construct V_tissue matrix from summary statistics and epigenomic tracks
}
