% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_vtrait.R
\name{make_vtrait}
\alias{make_vtrait}
\title{Construct V_trait matrix from PIP effects}
\usage{
make_vtrait(
  focal_trait,
  aux_traits,
  output_dir,
  pip_threshold = 0.01,
  aux_pip_thres = 0.5,
  aux_nbr_thres = 1,
  fine_mapped_files = NULL,
  full_matrix_path = NULL,
  subset = NULL
)
}
\arguments{
\item{focal_trait}{Name of the focal trait}

\item{aux_traits}{Character vector of auxiliary trait names}

\item{output_dir}{Path to output directory}

\item{pip_threshold}{Minimum PIP for the focal trait}

\item{aux_pip_thres}{Minimum PIP for auxiliary traits}

\item{aux_nbr_thres}{Minimum number of auxiliary traits with strong signal per SNP}

\item{fine_mapped_files}{Optional named list with paths to custom fine-mapped files. Names must match traits.}

\item{full_matrix_path}{Optional path to a precomputed full pleiotropy matrix (with columns RSID, *_pos, *_neg, A1)}

\item{subset}{Optional vector with a subset of traits from the precomputed full pleiotropy matrix (character vector with names)}
}
\value{
Path to the saved V_trait matrix
}
\description{
For a given focal trait and a list of auxiliary traits, this function computes the
directional PIP product for all SNPs shared across the traits, and saves the result.
If custom fine-mapped files are provided, they are converted to SQLite first.
}
